Summary: converter created for UTM to lat long coordinates system. Color maps are now registered in database for users.

+++ MapWebSite/MapWebSite.Model/Entities/ColorMap.cs
@@ -11,9 +11,9 @@ namespace MapWebSite.Model
     {
         public string Color { get; set; }
 
-        public int Left { get; set; }
+        public decimal Left { get; set; }
 
-        public int Right { get; set; }
+        public decimal Right { get; set; }
 
     }
